+++ Common/Models/StockModelDto.cs
@@ -16,5 +16,5 @@ public class StockModelDto
 public class StockValueIndex
 {
     public DateTime DateTime { get; set; } = DateTime.MinValue;
-    public double Value { get; set; } = 0;
+    public float Value { get; set; } = 0;
 }
